requirements.txt
python-dotenv>=1.0.1,<2
jinja2>=3.1.4,<4
playwright>=1.40,<2
